+++ src/main/kotlin/pl/pszklarska/pubversionchecker/util/Secrets.kt
@@ -1,0 +1,11 @@
+package pl.pszklarska.pubversionchecker.util
+
+/**
+ * Secrets are replaced by the CI's step. Ensure any changes in the filename or filepath are reflected there.
+ */
+
+class Secrets {
+    companion object {
+        const val sentryDsn = ""
+    }
+}